folder structure optimization
