+++ client/src/components/folders/createFolder.jsx
@@ -4,58 +4,34 @@ import ImageResize from "quill-image-resize-module-react";
 import { Link } from "react-router-dom";
 import ImageCompress from "quill-image-compress";
 import { convert as convertToText } from "html-to-text";
-import { formatDistance, set } from "date-fns";
-import { useParams } from "react-router-dom";
+import { formatDistance } from "date-fns";
 import parse from "html-react-parser";
 import { toast, ToastContainer } from "react-toastify";
 import { useNavigate } from "react-router-dom";
-import SelectLimit from "./helpers/selectLimit";
-import SelectOptions from "./helpers/selectOptions";
+import SelectLimit from "../helpers/selectLimit";
+import SelectOptions from "../helpers/selectOptions";
 import { Quill } from "react-quill";
 import "react-quill/dist/quill.snow.css";
-import "../styles/create.css";
+import "../../styles/create.css";
 import "react-toastify/dist/ReactToastify.css";
-import "../styles/allPages.css";
-import "../styles/folderCreate.css";
-import axiosInstance from "../utils/axiosConfig";
+import "../../styles/allPages.css";
+import "../../styles/folderCreate.css";
+import axiosInstance from "../../utils/axiosConfig";
 
-const EditFolder = () => {
+const CreateFolder = () => {
   const [title, setTitle] = useState("");
   const [description, setDescription] = useState("");
   const [allSets, setAllSets] = useState([]);
   const [setsChosen, setSetsChosen] = useState([]);
-  const [originalData, setOriginalData] = useState({});
   const [query, setQuery] = useState("");
   const [onlyPersonalSets, setOnlyPersonalSets] = useState(false);
   const [limit, setLimit] = useState(10);
   const [category, setCategory] = useState("");
   const [folderCategory, setFolderCategory] = useState("");
-
   Quill.register("modules/imageResize", ImageResize);
   Quill.register("modules/imageCompress", ImageCompress);
-  const { id } = useParams();
-  const navigate = useNavigate();
 
-  const getFolder = async () => {
-    axiosInstance.get(`/folder/${id}`).then((res) => {
-      setTitle(res.data[0].title);
-      setDescription(res.data[0].description);
-      //get set ids from the array of res.data and set it to setsChosen
-      let setIds = [];
-      console.log(res.data);
-      res.data.forEach((el) => {
-        setIds.push(el.set_id);
-      });
-      setFolderCategory(res.data[0].category);
-      setOriginalData({
-        title: res.data[0].title,
-        description: res.data[0].description,
-        sets: setIds,
-        category: res.data[0].category,
-      });
-      setSetsChosen(setIds);
-    });
-  };
+  const navigate = useNavigate();
 
   const getSets = async () => {
     axiosInstance
@@ -76,7 +52,7 @@ const EditFolder = () => {
       setSetsChosen([...setsChosen, id]);
     }
   };
-  const updateFolder = async () => {
+  const createFolder = async () => {
     if (setsChosen.length > 2000) {
       toast.error("More than 2000 sets are not allowed in a folder");
       return;
@@ -99,32 +75,21 @@ const EditFolder = () => {
     }
     if (setsChosen.length < 1) {
       if (
-        !confirm("Are you sure you want to update a folder without any sets?")
+        !confirm("Are you sure you want to create a folder without any sets?")
       ) {
         return;
       }
     }
-    if (
-      title === originalData.title &&
-      description === originalData.description &&
-      setsChosen.length === originalData.sets.length &&
-      setsChosen.every((val, index) => val === originalData.sets[index]) &&
-      folderCategory === originalData.folderCategory
-    ) {
-      toast.error("You didn't make any changes");
-      return;
-    }
     axiosInstance
-      .post("/folder/update", {
-        title,
-        description,
-        setsChosen,
-        folder_id: id,
-        folderCategory,
+      .post("/folders/create/", {
+        title: title,
+        description: description,
+        sets: setsChosen,
+        category: folderCategory,
       })
       .then((res) => {
-        toast.success("Folder edited successfully");
-        navigate(`/folder/${id}`);
+        toast.success("Folder created successfully");
+        navigate(`/folder/${res.data.folder_id}`);
       })
       .catch((err) => {
         toast.error("Something went wrong");
@@ -132,10 +97,9 @@ const EditFolder = () => {
   };
 
   useEffect(() => {
-    getFolder();
     getSets();
   }, []);
-  useEffect(() => console.log(originalData.category), [originalData.category]);
+
   return (
     <section>
       <ToastContainer
@@ -222,22 +186,13 @@ const EditFolder = () => {
               },
             }}
           />
-          <div style={{margin:"1vmax"}}>
-          {originalData?.category ? (
-            <SelectOptions
-              setCategory={setFolderCategory}
-              initialState={originalData.category}
-            />
-          ) : (
-            ""
-          )}
+          <div style={{ margin: "1vmax" }}>
+            <SelectOptions setCategory={setFolderCategory} />
           </div>
           <div id="setsChosenContainer">
             <center>
               {" "}
-              <p>
-                Your preselected sets are already here, you can always add new
-              </p>
+              <p>You can also add some of your sets in the folder</p>
             </center>
             <div id="searchContainer">
               <center>
@@ -264,7 +219,7 @@ const EditFolder = () => {
                   <div
                     className={"individualSet"}
                     style={{
-                      maxWidth: "20vw",
+                      maxWidth: "20vmax",
                       boxShadow: setsChosen.includes(el.set_id)
                         ? "0 19px 38px rgba(0,0,0,0.30), 0 15px 12px rgba(0,0,0,0.22)"
                         : "0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)",
@@ -317,14 +272,14 @@ const EditFolder = () => {
                   boxShadow:
                     "0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24)",
                 }}
-                onClick={updateFolder}
+                onClick={createFolder}
               >
-                Update folder with
+                Create folder{" "}
                 {setsChosen.length
-                  ? ` ${setsChosen.length} set${
+                  ? `with ${setsChosen.length} set${
                       setsChosen.length === 1 ? "" : "s"
                     }`
-                  : " no sets"}
+                  : "no sets"}
               </button>
             </center>
           </div>
@@ -334,4 +289,4 @@ const EditFolder = () => {
   );
 };
 
-export default EditFolder;
+export default CreateFolder;
